requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
groq==0.5.0
google-generativeai==0.4.0
apscheduler==3.10.4
python-dotenv==1.0.1
lxml==5.1.0
